Migrate to fastify session for user mutations
